Feature: update verovio toolkit
The new version of verovio supports rendering of chords and grace notes.
The preparation of the XML code was adjusted accordingly to prevent
chords from being escaped to layers as before. A CSS rule for
svg:not(:root) had to be overwritten in order to enable correct display
of the SVG.

closes #44